automation-service/requirements.txt
alembic==1.10.2
aniso8601==9.0.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
cachelib==0.9.0
certifi==2023.5.7
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.3.0
eventlet==0.33.3
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
flask-paginate==2022.1.8
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mistune==3.0.1
packaging==23.1
passlib==1.7.4
pip-autoremove==0.10.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
python-decouple==3.8
python-dotenv==1.0.0
python-engineio==4.4.1
pytz==2023.3
redis==4.5.5
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.6.3
urllib3==1.26.13
Werkzeug==2.3.6
zipp==3.15.0
